Map = screen size * 2
